rolesScope functions moved to multiplexer
